[Feature] Add CalculationResults component, split calculator into submodules
