+++ src/App.css
@@ -26,3 +26,20 @@
 .cfn-logo{
   height: 3.75em;
 }
+
+.calculator-container{
+  min-height: 100vh;
+}
+
+.price-form{
+  width: 80%;
+  margin: 0 auto;
+}
+
+.navbar{
+  height: 7em;
+}
+
+.navbar-right{
+  padding-top: 25px;
+}
