+++ src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import './App.css';
 import { BrowserRouter, Route } from 'react-router-dom';
 import NavigationBar from './components/Navbar';
-import Calculator from './components/Calculator';
+import Calculator from './components/calculator/Calculator';
 
 class App extends Component {
   render() {
